move extensionsign pkg
