+++ extensionsign/sigmanifest.go
