+++ extensionsign/sigmanifest_test.go
