+++ extensionsign/sigzip.go
